spi: qup: Wait for QUP to complete
This patch changes the completion to occur when the QUP
controller specifies that all of the bytes have been
written/read.  This removes the timing issue that can
occur when the DMA completes the copy to the FIFO, but
the QUP has not sent out the bytes.  When this happens,
the QUP may not send out all the bytes in the FIFO before
it is placed back in the reset state.

CC: Sarthak Kukreti <skukreti@codeaurora.org>
CC: Bryan Freed <bfreed@chromium.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>
Signed-off-by: Andy Gross <agross@codeaurora.org>